Additional changes
Fixed issues with timed executions
Documentation additions
Consolidated common routines.
Fixed PKI issue

+++ src/main/java/com/elastic/support/util/SystemUtils.java
@@ -2,6 +2,7 @@ package com.elastic.support.util;
 
 import com.elastic.support.Constants;
 import com.elastic.support.diagnostics.DiagnosticException;
+import com.elastic.support.diagnostics.commands.GenerateManifest;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.logging.log4j.LogManager;
@@ -25,6 +26,12 @@ public class SystemUtils {
         logger.info(Constants.CONSOLE,  "Exiting...");
         System.exit(0);
     }
+
+    public static String getToolVersion() {
+        String ver = GenerateManifest.class.getPackage().getImplementationVersion();
+        return (ver != null) ? ver : Constants.runningInIde;
+    }
+
 
     public static void writeToFile(String content, String dest) {
         try {
@@ -54,11 +61,14 @@ public class SystemUtils {
     public static void nukeDirectory(String dir){
         try {
             File tmp = new File(dir);
+            if(!tmp.exists()){
+                return;
+            }
             tmp.setWritable(true, false);
             FileUtils.deleteDirectory(tmp);
             logger.info(Constants.CONSOLE,  "Deleted directory: {}.", dir);
         } catch (IOException e) {
-            logger.error(Constants.CONSOLE, "Delete of directory:{} failed. Usually this indicates a permission issue", dir, e);
+            logger.error(Constants.CONSOLE, "Delete of directory:{} failed. This may indicate a permission issue", dir, e);
         }
     }
